+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Rename subcommand `TransactionSubmit` to `TransactionSignAndSubmit`.
+- Add subcommand `TransactionSubmit` to submit a signed transaction on chain.
+- Add subcommand `TransactionAddSignature` to add a signature to a partially-signed transaction.
+- Add optional `--out` flag to all transaction-creating commands to output a partially-singed transaction to a file.
 - Update GHC version to 9.6.4 (lts-22.9).
 - Update Rust version to 1.73.
 - Support protocol version 7.
@@ -58,7 +62,7 @@
 - Add round and epoch to the output of `raw GetBlockInfo`.
 - Add round and epoch to the output of `block show` when they are present.
 - Print "Block time" instead of "Slot time" in the output of `block show`.
-- In the output of `consensus show-parameters`, only print election difficulty when present. 
+- In the output of `consensus show-parameters`, only print election difficulty when present.
 
 ## 5.2.0
 
@@ -121,10 +125,10 @@
 
 - Add support of contract schema V3.
 	  - V3 schemas offer the same options as V2, but also optionally includes a schema for contract events.
-	  - `transaction status` now displays contract events, and a schema can be provided with `--schema`, which 
+	  - `transaction status` now displays contract events, and a schema can be provided with `--schema`, which
 	    will be used to parse the contract events. By default events are parsed with the schema embedded in the
       contract, if present.
-	  - This enables concordium-client to interact with contracts and schemas 
+	  - This enables concordium-client to interact with contracts and schemas
 	    using `concordium-std` version 5.
 - Improved formatting of `transaction status` output using contract schemas if
   they are available for displaying contract events.